Fix formatting

+++ src/cascadia/TerminalControl/TermControl.h
@@ -93,7 +93,7 @@ namespace winrt::Microsoft::Terminal::TerminalControl::implementation
         std::optional<int> _lastScrollOffset;
 
         // Auto scroll occurs when user, while selecting, drags cursor outside viewport. View is then scrolled to 'follow' the cursor.
-        float _autoScrollVelocity;
+        double _autoScrollVelocity;
         std::optional<Windows::UI::Input::PointerPoint> _autoScrollingPointerPoint;
         Windows::UI::Xaml::DispatcherTimer _autoScrollTimer;
 		static constexpr double _AutoScrollTimerInterval = 1.0 / 30.0;
